feat: start work on decomposition of file upload component
