+++ src/components/mail-sender/mail-form.jsx
@@ -1,29 +1,13 @@
 import React from 'react';
 import { Formik } from 'formik';
-import useForm from '../../hooks/use-form';
-import DropZone from './drop-zone';
-import AttachmentItem from './attachment-item';
-import attachmentIcon from '../../assets/images/attachment.svg';
 import validationSchema from '../../utils/form-validation';
-import { createRef } from 'react/cjs/react.development';
+import Attachments from '../attachments/attachments';
 import '../../scss/mail-sender.scss';
 
 function MailForm() {
-  const fileRef = createRef();
-  const dropZoneRef = createRef();
-  const {
-    preventDefaults,
-    handleDrop,
-    handleRemoveAttachment,
-    handleDragOver,
-    handleDragLeave,
-    handleSelectFile,
-    handleSelectFileChange,
-    attachmentList
-  } = useForm(fileRef, dropZoneRef);
   return (
     <div className="main">
-      <div className="container" onDragOver={handleDragOver}>
+      <div className="container" onDragOver={() => {}}>
         <div className="main__title">Отправлялка сообщений</div>
         <Formik
           initialValues={{
@@ -159,63 +143,7 @@ function MailForm() {
                     </div>
                   </div>
                 </div>
-                <div className="main__row">
-                  <div className="attachment">
-                    {attachmentList.map(({ attachment, id }) => (
-                      <AttachmentItem
-                        key={id}
-                        id={id}
-                        attachment={attachment}
-                        onRemoveAttachment={handleRemoveAttachment}
-                      />
-                    ))}
-                  </div>
-                </div>
-                <div className="main__row">
-                  <div
-                    className="upload"
-                    onDragEnter={preventDefaults}
-                    onDragOver={preventDefaults}
-                    onDragLeave={preventDefaults}
-                    onDrop={handleDrop}>
-                    <div className="upload__item">
-                      <div className="upload__row">
-                        <div className="upload__image">
-                          <img src={attachmentIcon} alt="" />
-                        </div>
-                        <input
-                          type="file"
-                          ref={fileRef}
-                          onChange={handleSelectFileChange}
-                          className="upload__file"
-                          multiple
-                        />
-                        <div className="upload__title hover" onClick={handleSelectFile}>
-                          Прикрепить файл
-                        </div>
-                      </div>
-                      <div className="upload__row">
-                        {/* <div className="upload__limit">
-                          {attachmentsLimitInfo.noOfAttachmentsExceededLimit
-                            ? `Максимальный размер одного файла не должен превышать 5 МБ`
-                            : ''}
-                          <br />
-                          {attachmentsLimitInfo.totalAttachmentsSizeExceeded
-                            ? `Объем всех вложении не должен превышать 20 МБ`
-                            : ''}
-                        </div> */}
-                      </div>
-                    </div>
-                  </div>
-                </div>
-                <DropZone
-                  dropZoneRef={dropZoneRef}
-                  preventDefaults={preventDefaults}
-                  onDragEnter={preventDefaults}
-                  onDragOver={preventDefaults}
-                  onDragLeave={handleDragLeave}
-                  onDrop={handleDrop}
-                />
+                <Attachments />
               </form>
             );
           }}
